+++ example/main.dart
@@ -30,27 +30,27 @@ class MyHomePage extends StatelessWidget {
           itemCount: 10,
           itemBuilder: (context, index) {
             return ExpandableCard(
-              textButtonActionFirst: 'Close',
-              textButtonActionSecond: 'Details',
+              collapsedButtonText: 'Details',
+              expandedButtonText: 'Close',
               title: 'Fried Rice',
               description: 'Confirmed Order',
-              button2Value: 'Buy Now',
+              actionButtonText: 'Buy Now',
               sectionRowCount: 3,
               sectionRowTitles: const ['Review', 'Order', 'Shipping'],
               totalText: 3,
               backgroundColor: Colors.white,
               elevation: 4.0,
-              button2Elevation: 5.0,
-              button2Color: Colors.blue,
-              button1TextColor: Colors.black,
-              button2BorderRadius: 5.0,
+              actionButtonElevation: 5.0,
+              actionButtonColor: Colors.blue,
+              expandCollapseButtonTextColor: Colors.black,
+              actionButtonBorderRadius: 5.0,
               cardBorderRadius: 10,
               sectionRowData: const {
                 'Review': ['Good portion size', 'Taste good', 'Overall good'],
                 'Order': ['Fried Rice', '1', 'RM 10.00'],
                 'Shipping': ['Street 1', 'City 1', '12345'],
               },
-              onPressedButton2: () {
+              onActionButtonPressed: () {
                 showMovePage(context);
               },
             );