+++ lib/src/components/title_description_section.dart
@@ -6,13 +6,13 @@ class TitleDescriptionSection extends StatelessWidget {
   final String? title;
 
   /// The description to display.
-  final String? deskripsi;
+  final String? description;
 
   /// Creates a title and description section.
   const TitleDescriptionSection({
     Key? key,
     this.title,
-    this.deskripsi,
+    this.description,
   }) : super(key: key);
 
   @override
@@ -34,7 +34,7 @@ class TitleDescriptionSection extends StatelessWidget {
         // Description text widget with padding
         Padding(
           padding: const EdgeInsets.only(left: 12, top: 10),
-          child: Text(deskripsi ??
+          child: Text(description ??
               ''), // Display description text or empty string if null
         ),
         const SizedBox(height: 16), // Spacer widget for additional spacing